+++ src/__tests__/mocks/notifications/subscription-activated.mock.ts
@@ -223,7 +223,6 @@ export const SubscriptionActivatedMockExpectation = {
     scheduledChange: null,
     startedAt: '2023-08-11T08:07:35.449123Z',
     status: 'active',
-    transactionId: 'txn_01h8bxpvx398a7zbawb77y0kp5',
     updatedAt: '2023-08-11T08:07:36.892822Z',
   },
   eventId: 'evt_01h7ht60mmw6d4sf4h38g3t4yq',