+++ src/notifications/events/subscription/subscription-past-due-event.ts
@@ -12,7 +12,7 @@ import { type ISubscriptionNotificationResponse } from '../../types/index.js';
 
 export class SubscriptionPastDueEvent extends Event {
   public override readonly eventType = EventName.SubscriptionPastDue;
-  public override readonly data: Omit<SubscriptionNotification, 'transactionId'>;
+  public override readonly data: SubscriptionNotification;
 
   constructor(response: IEventsResponse<ISubscriptionNotificationResponse>) {
     super(response);