+++ src/notifications/types/subscription/index.ts
@@ -4,6 +4,7 @@
  *  Changes may be overwritten as part of auto-generation.
  */
 
+export * from './subscription-created-notification-response.js';
 export * from './subscription-discount-notification-response.js';
 export * from './subscription-time-period-notification-response.js';
 export * from './subscription-scheduled-change-notification-response.js';